:sparkles: separeted component file
